build.sh mac vs linux

+++ conda/ta-lib/build.sh
@@ -1,20 +1,23 @@
 #!/bin/bash
-wget https://downloads.sourceforge.net/project/ta-lib/ta-lib/0.4.0/ta-lib-0.4.0-src.tar.gz &&
-tar xvfz ta-lib-0.4.0-src.tar.gz &&
-sudo apt-get update &&
-sudo apt-get install gcc-multilib g++-multilib &&
-echo "Installing TA-Lib in $PREFIX" &&
-ls ta-lib &&
-pushd ta-lib &&
-./configure --prefix=$PREFIX &&
-make &&
-make install &&
-popd &&
-rm ta-lib-0.4.0-src.tar.gz &&
-rm -r ta-lib &&
-
-export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$PREFIX/lib &&
-export TA_INCLUDE_PATH=$PREFIX/include &&
-export TA_LIBRARY_PATH=$PREFIX/lib &&
-python setup.py build &&
-python setup.py install --prefix=$PREFIX
+if [[ "$OSTYPE" == "linux-gnu" ]]; then
+  wget https://downloads.sourceforge.net/project/ta-lib/ta-lib/0.4.0/ta-lib-0.4.0-src.tar.gz &&
+    tar xvfz ta-lib-0.4.0-src.tar.gz &&
+    sudo apt-get update &&
+    sudo apt-get install build-essential gcc-multilib g++-multilib &&
+    cd ta-lib &&
+    ./configure --prefix=$PREFIX &&
+    make &&
+    make install &&
+    cd .. &&
+    rm -rf ta-lib &&
+    rm ta-lib-0.4.0-src.tar.gz &&
+    export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$PREFIX/lib &&
+    export TA_INCLUDE_PATH=$PREFIX/include &&
+    export TA_LIBRARY_PATH=$PREFIX/lib &&
+    python setup.py build &&
+    python setup.py install --prefix=$PREFIX
+elif [[ "$OSTYPE" == "darwin"* ]]; then
+  brew install ta-lib &&
+    python setup.py build &&
+    python setup.py install
+fi
